LPD-95990 Move source under a js folder
