+++ modules/apps/frontend-js/frontend-js-charts-web/src/main/resources/META-INF/resources/js/chart_state_wrapper/ChartState.tsx
